update watch history section
